fix(messaging): address PostgreSQL + SqlServer storage review findings (#200)
* fix(messaging): wrap retry queries in transaction, handle IDbContextTransaction, parameterize lock TTL

- _GetMessagesOfNeedRetryAsync: FOR UPDATE SKIP LOCKED (PG) / UPDLOCK (SQL) now
  runs inside explicit transaction so row locks are held until commit
- _ChangeMessageStateAsync: handle IDbContextTransaction branch matching
  StoreMessageAsync pattern — prevents silent execution outside caller's transaction
- RenewLockAsync: parameterize TTL seconds instead of string interpolation —
  fixes culture-sensitive formatting and plan cache pollution

* fix(messaging): change monitoring count types from int to long

PostgreSQL COUNT() returns bigint; reading with GetInt32 throws
InvalidCastException when count exceeds Int32.MaxValue. Changed
StatisticsView properties, IMonitoringApi count methods, provider
implementations, and DbConnectionExtensions.ExecuteScalarAsync to use
long/Int64 consistently.

BREAKING CHANGE: StatisticsView count properties and IMonitoringApi
count method return types changed from int to long.

* fix(postgres): use TIMESTAMPTZ, wrap MessageId migration in DO block

- Changed all TIMESTAMP columns to TIMESTAMPTZ in CREATE TABLE DDL for
  published, received, and lock tables
- Added idempotent migration: existing TIMESTAMP columns are converted
  to TIMESTAMPTZ via ALTER COLUMN TYPE ... USING ... AT TIME ZONE 'UTC'
- Consolidated MessageId migration into a single DO block that handles
  both fresh-add and nullable→NOT NULL cases atomically
- Prevents timezone confusion when server timezone ≠ UTC

* perf(postgres): use array param, window function, time range predicate

- ChangePublishStateToDelayedAsync: replace N-parameter IN clause with
  = ANY(@Ids) array parameter — enables plan caching, removes 65535 param limit
- GetMessagesAsync: merge COUNT+SELECT into single query using
  COUNT(*) OVER() window function — eliminates duplicate round-trip
- _GetHourlyTimelineStats: push time range predicate (Added >= @MinAdded)
  into base query so existing indexes can prune rows before aggregation

* refactor(messaging): rename fields, remove N'' prefix, fix README

- Rename _pubName/_recName/_lockName to _publishedTable/_receivedTable/
  _lockTable in both PostgreSql and SqlServer providers
- Remove SQL Server N'' string prefix from PostgreSQL GetStatisticsAsync
- Fix PostgreSQL README: remove nonexistent TableNamePrefix option,
  correct table names to schema-based naming

* test(messaging): strengthen harness assertions, add monitoring tests

- DataStorageTestsBase: strengthen assertions for store_message_with_transaction,
  handle_failed_message_state (verify message in retry results), delete assertions
  changed from >= 0 to == 1
- Add PostgreSqlMonitoringTest: integration tests for GetStatisticsAsync,
  count methods, GetPublishedMessageAsync with seeded data

* chore(postgres): remove dead migration blocks — no existing consumers

Greenfield project with no deployed databases. The MessageId backfill,
nullable→NOT NULL migration, and TIMESTAMP→TIMESTAMPTZ ALTER blocks
were dead code. CREATE TABLE already uses the correct types.

* fix(messaging): read COUNT(*) OVER() as Int64, clamp to int for IndexPage

* fix(messaging): address code review P1/P2 findings

- PostgreSqlStorageInitializer: use DateTimeKind.Utc for DateTime.MinValue
  on TIMESTAMPTZ column — Npgsql 7+ rejects Kind=Unspecified
- SqlServerDataStorage: cast TtlSeconds to int for DATEADD — SQL Server
  rejects double/float for integer datepart argument
- PostgreSqlMonitoringApi._GetTimelineStats: GetInt64 for COUNT() bigint
  return type (same fix already applied to GetStatisticsAsync)
- PostgreSqlMonitoringApi: forward CancellationToken to _GetNumberOfMessage
  from all four count methods
- Remove unnecessary await Task.CompletedTask in test

* perf(sqlserver): push time range predicate into _GetTimelineStats

Match PostgreSQL fix — filter by Added >= @MinAdded before GROUP BY
so indexes can prune rows instead of scanning the full table.

* fix(sqlserver): address review findings — dead migration, null safety, sealed, COUNT OVER

- Remove dead MessageId migration block (greenfield, no consumers)
- StoreMessageAsync: throw InvalidOperationException for unsupported
  transaction types instead of NullReferenceException (both providers)
- Seal SqlServerMonitoringApi
- Fix NoopTransaction discard pattern
- Port COUNT(*) OVER() single-query pattern to SqlServer GetMessagesAsync

* fix(postgres): DateTimeKind.Utc in ReleaseLockAsync, forward CancellationToken, remove Lower()

- ReleaseLockAsync: DateTime.MinValue needs Kind=Utc for TIMESTAMPTZ column
  (same fix already applied to InitializeAsync, missed here)
- HourlySucceededJobs/HourlyFailedJobs: forward cancellationToken to
  _GetHourlyTimelineStats (SqlServer already did this)
- _GetNumberOfMessage: remove Lower() wrapper on StatusName — values are
  controlled enum strings, bare equality enables partial index use

* fix(postgres): remove Lower() filters, add comprehensive monitoring tests

- Remove Lower() from GetMessagesAsync StatusName/Name/Group filters —
  values are controlled, bare equality enables index use
- Add TIMESTAMPTZ schema assertion test (information_schema verification)
- Add GetMessagesAsync pagination and status filter tests
- Add HourlySucceededJobs/HourlyFailedJobs tests
- Add GetReceivedMessageAsync and null-return tests
- Add StoreMessageAsync with real NpgsqlTransaction test
- Assert all 5 StatisticsView fields (was only checking 2)

* fix(sqlserver): align totalCount to long, add ConfigureAwait, fix TTL cast

- GetMessagesAsync: totalCount int→long + GetInt64 + Math.Min clamp,
  matching PostgreSQL implementation
- GetStatisticsAsync: add missing ConfigureAwait(false) on ReadAsync
- RenewLockAsync: (int)→(long) cast to avoid silent truncation

* test(postgres): add rollback, unsupported tx, empty page, renew lock tests

- should_rollback_transaction_and_not_persist_message: verify rolled-back
  StoreMessageAsync leaves no row
- should_throw_for_unsupported_transaction_type: verify InvalidOperationException
  for non-DbTransaction/IDbContextTransaction objects
- should_return_empty_page_when_no_messages_match: verify TotalItems=0
  from COUNT(*) OVER() path with empty result set
- should_renew_lock_and_advance_last_lock_time: read LastLockTime before
  and after RenewLockAsync, assert it advanced
- Fix lock test dispose: use timezone-aware literal for TIMESTAMPTZ column

* fix(sqlserver): align ReleaseLockAsync sentinel and timeline count type with PG

- ReleaseLockAsync: use 1900-01-01 UTC sentinel matching InitializeAsync
  (was DateTime.MinValue with Unspecified Kind)
- _GetTimelineStats: read COUNT with GetInt64 then cast to int, matching
  PostgreSQL's safer pattern

+++ src/Headless.Messaging.Core/Monitoring/IMonitoringApi.cs
@@ -18,13 +18,13 @@ public interface IMonitoringApi
         CancellationToken cancellationToken = default
     );
 
-    ValueTask<int> PublishedFailedCount(CancellationToken cancellationToken = default);
+    ValueTask<long> PublishedFailedCount(CancellationToken cancellationToken = default);
 
-    ValueTask<int> PublishedSucceededCount(CancellationToken cancellationToken = default);
+    ValueTask<long> PublishedSucceededCount(CancellationToken cancellationToken = default);
 
-    ValueTask<int> ReceivedFailedCount(CancellationToken cancellationToken = default);
+    ValueTask<long> ReceivedFailedCount(CancellationToken cancellationToken = default);
 
-    ValueTask<int> ReceivedSucceededCount(CancellationToken cancellationToken = default);
+    ValueTask<long> ReceivedSucceededCount(CancellationToken cancellationToken = default);
 
     ValueTask<Dictionary<DateTime, int>> HourlySucceededJobs(
         MessageType type,
